split test into 2 files
